+++ src/TitleDetails/TitleInformation/TitleInformation.test.js
@@ -1,9 +1,7 @@
-import React from 'react';
-import { MemoryRouter } from 'react-router-dom';
-import { render, cleanup } from '@folio/jest-config-stripes/testing-library/react';
 import { IntlProvider } from 'react-intl';
+import { MemoryRouter } from 'react-router-dom';
 
-import '@folio/stripes-acq-components/test/jest/__mock__';
+import { render, cleanup } from '@folio/jest-config-stripes/testing-library/react';
 
 import TitleInformation from './TitleInformation';
 
@@ -11,6 +9,8 @@ const renderTitleInformation = (titleProp) => (render(
   <IntlProvider locale="en">
     <MemoryRouter>
       <TitleInformation
+        claimingActive={titleProp.claimingActive}
+        claimingInterval={titleProp.claimingInterval}
         contributors={titleProp.contributors}
         edition={titleProp.edition}
         productIds={titleProp.productIds}
@@ -25,6 +25,8 @@ const renderTitleInformation = (titleProp) => (render(
 ));
 
 const title = {
+  claimingActive: true,
+  claimingInterval: 42,
   contributors: [],
   edition: 'First edition',
   productIds: [],
@@ -47,5 +49,7 @@ describe('Given Title information', () => {
     expect(getByText(title.subscriptionFrom)).toBeDefined();
     expect(getByText(title.subscriptionInterval.toString())).toBeDefined();
     expect(getByText(title.subscriptionTo)).toBeDefined();
+    expect(getByText('ui-receiving.title.claimingActive')).toBeInTheDocument();
+    expect(getByText(title.claimingInterval)).toBeInTheDocument();
   });
 });